Fix a bug that a rollback point with zero objects causes an error.
This should rarely happen, but an observed case is the following.
The freebsd-update server distributes a file in /usr/src to a system
without src.  In the installation phase, the file is not installed
because the target directory does not exist.  In the next fetch phase,
there is no real updates, but the file alone is distributed and failed
to be installed again.  The rollback point generated here refers only
that object.  When that object is referred by another new rollback point,
the former point loses its ownership of the object.